pyproject.toml
[project]
name = "autoclipper"
version = "0.1.0"
description = "AutoClipper - Video and Audio Clipping Tool"
requires-python = ">=3.12.6"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "aliyun-python-sdk-core==2.16.0",
    "aliyun-python-sdk-kms==2.16.5",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "audioread==3.1.0",
    "azure-identity>=1.25.1",
    "brotli==1.2.0",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "crcmod==1.7",
    "cryptography==46.0.3",
    "curl-cffi==0.13.0",
    "decorator==4.4.2",
    "distro==1.9.0",
    "editdistance==0.8.1",
    "fastapi==0.122.0",
    "ffmpy==1.0.0",
    "filelock==3.20.0",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "funasr==1.2.7",
    "google-generativeai>=0.8.0",
    "gradio==6.0.1",
    "gradio-client==2.0.0",
    "groovy==0.1.2",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==1.1.5",
    "hydra-core==1.3.2",
    "idna==3.11",
    "imageio==2.37.2",
    "imageio-ffmpeg==0.6.0",
    "jaconv==0.4.0",
    "jamo==0.4.1",
    "jieba==0.42.1",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "jmespath==0.10.0",
    "joblib==1.5.2",
    "kaldiio==2.18.1",
    "lazy-loader==0.4",
    "librosa==0.11.0",
    "llvmlite==0.45.1",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "modelscope==1.32.0",
    "moviepy==1.0.3",
    "mpmath==1.3.0",
    "msgpack==1.1.2",
    "multidict==6.7.0",
    "nest-asyncio==1.6.0",
    "networkx==3.6",
    "numba==0.62.1",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "openai==2.8.1",
    "orjson==3.11.4",
    "oss2==2.19.1",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==12.0.0",
    "platformdirs==4.5.0",
    "pooch==1.8.2",
    "proglog==0.1.12",
    "propcache==0.4.1",
    "protobuf>=4.25.0,<6.0.0",
    "pycparser==2.23",
    "pycryptodome==3.23.0",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydub==0.25.1",
    "pygments==2.19.2",
    "pynndescent==0.5.13",
    "pyright>=1.1.407",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.0.0",
    "python-multipart==0.0.20",
    "pytorch-wpe==0.0.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "rich==14.2.0",
    "ruff==0.14.6",
    "safehttpx==0.1.7",
    "scikit-learn==1.7.2",
    "scipy==1.16.3",
    "semantic-version==2.10.0",
    "sentencepiece==0.2.1",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soundfile==0.13.1",
    "soxr==1.0.0",
    "starlette==0.50.0",
    "sympy==1.14.0",
    "tensorboardx==2.6.4",
    "threadpoolctl==3.6.0",
    "tomlkit==0.13.3",
    "torch==2.9.1",
    "torch-complex==0.4.4",
    "torchaudio==2.9.1",
    "tqdm==4.67.1",
    "typer==0.20.0",
    "typer-slim==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "umap-learn==0.5.9.post2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "websocket-client==1.9.0",
    "yarl==1.22.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["funclip"]

# Ruff setting
[tool.ruff]
# Python version
target-version = "py38"
line-length = 100
indent-width = 4

exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# ruff lint
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

# 
ignore = [
    "E501", 
    "B008", 
    "C901",
]


[tool.ruff.lint.isort]
known-first-party = ["funclip"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"


[tool.pyright]
include = ["funclip"]
exclude = ["**/node_modules", "**/__pycache__", "**/.*"]
reportMissingImports = true
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportMissingParameterType = false
pythonVersion = "3.8"
typeCheckingMode = "basic"


